microservices/micro02/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
PyMySQL==1.1.0
redis==5.0.1
python-dotenv==1.0.0
firebase-admin==6.4.0
